Add a devkit, rudimentary plugin support and a sdk.
The devkit is a plugin which can be enabled in the config of the daemon.
Using this tool, it is possible to control the console over HTTP,
provided one has the required access tokens to issue commands.

To aid developmemt, the devkit is also able to export the / directory
over NFS to the developer machine. Using this approach, development will
become significantly easier and faster. Scripts to make this easier are
going to follow in the next few commits.